###What is the purpose of the change Related issues: #655 Enhance Linkis JobHistory Module to support task cache, label persistence and manager of job history.
### Brief change log
The module contains follow changes:
1. Add cache of task in querying
2. Add persistence of label
3. Add admin of job history
