+++ linkis-public-enhancements/linkis-publicservice/linkis-jobhistory/src/main/java/com/webank/wedatasphere/linkis/jobhistory/cache/QueryCacheManager.java
@@ -1,0 +1,12 @@
+package com.webank.wedatasphere.linkis.jobhistory.cache;
+
+import com.webank.wedatasphere.linkis.jobhistory.cache.impl.UserTaskResultCache;
+
+public interface QueryCacheManager {
+
+    UserTaskResultCache getCache(String user, String engineType);
+
+    void cleanAll();
+
+    void refreshAll();
+}
