+++ linkis-public-enhancements/linkis-publicservice/linkis-jobhistory/src/main/scala/com/webank/wedatasphere/linkis/jobhistory/conf/JobhistoryConfiguration.scala
@@ -1,0 +1,8 @@
+package com.webank.wedatasphere.linkis.jobhistory.conf
+
+import com.webank.wedatasphere.linkis.common.conf.CommonVars
+
+object JobhistoryConfiguration {
+  val JOB_HISTORY_ADMIN = CommonVars("wds.linkis.jobhistory.admin", "enjoyyin,alexyang,allenlliu,chaogefeng,cooperyang,johnnwang,shanhuang,leeli,neiljianliu,kakacai,neilliao")
+  val JOB_HISTORY_SAFE_TRIGGER = CommonVars("wds.linkis.jobhistory.safe.trigger", true).getValue
+}
